fix code review issues
