+++ src/analytics/private/sql-def.ts
@@ -98,67 +98,80 @@ export const reportingViewsDef: SQLViewDef[] = [
     viewName: CLICKSTREAM_DEVICE_CRASH_RATE,
   },
   {
-    viewName: CLICKSTREAM_ACQUISITION_COUNTRY_NEW_USER_SP,
+    viewName: CLICKSTREAM_ACQUISITION_COUNTRY_NEW_USER,
+    spName: CLICKSTREAM_ACQUISITION_COUNTRY_NEW_USER_SP,
     type: 'sp',
     scheduleRefresh: 'true',
   },
   {
-    viewName: CLICKSTREAM_ACQUISITION_DAY_TRAFFIC_SOURCE_USER_SP,
+    viewName: CLICKSTREAM_ACQUISITION_DAY_TRAFFIC_SOURCE_USER,
+    spName: CLICKSTREAM_ACQUISITION_DAY_TRAFFIC_SOURCE_USER_SP,
     type: 'sp',
     scheduleRefresh: 'true',
   },
   {
-    viewName: CLICKSTREAM_ACQUISITION_DAY_USER_ACQUISITION_SP,
+    viewName: CLICKSTREAM_ACQUISITION_DAY_USER_ACQUISITION,
+    spName: CLICKSTREAM_ACQUISITION_DAY_USER_ACQUISITION_SP,
     type: 'sp',
     scheduleRefresh: 'true',
   },
   {
-    viewName: CLICKSTREAM_ENGAGEMENT_DAY_USER_VIEW_SP,
+    viewName: CLICKSTREAM_ENGAGEMENT_DAY_USER_VIEW,
+    spName: CLICKSTREAM_ENGAGEMENT_DAY_USER_VIEW_SP,
     type: 'sp',
     scheduleRefresh: 'true',
   },
   {
-    viewName: CLICKSTREAM_ENGAGEMENT_ENTRANCE_SP,
+    viewName: CLICKSTREAM_ENGAGEMENT_ENTRANCE,
+    spName: CLICKSTREAM_ENGAGEMENT_ENTRANCE_SP,
     type: 'sp',
     scheduleRefresh: 'true',
   },
   {
-    viewName: CLICKSTREAM_ENGAGEMENT_EXIT_SP,
+    viewName: CLICKSTREAM_ENGAGEMENT_EXIT,
+    spName: CLICKSTREAM_ENGAGEMENT_EXIT_SP,
     type: 'sp',
     scheduleRefresh: 'true',
   },
   {
-    viewName: CLICKSTREAM_ENGAGEMENT_KPI_SP,
+    viewName: CLICKSTREAM_ENGAGEMENT_KPI,
+    spName: CLICKSTREAM_ENGAGEMENT_KPI_SP,
     type: 'sp',
     scheduleRefresh: 'true',
   },
   {
-    viewName: CLICKSTREAM_ENGAGEMENT_PAGE_SCREEN_VIEW_DETAIL_SP,
+    viewName: CLICKSTREAM_ENGAGEMENT_PAGE_SCREEN_VIEW_DETAIL,
+    spName: CLICKSTREAM_ENGAGEMENT_PAGE_SCREEN_VIEW_DETAIL_SP,
     type: 'sp',
     scheduleRefresh: 'true',
   },
   {
-    viewName: CLICKSTREAM_ENGAGEMENT_PAGE_SCREEN_VIEW_SP,
+    viewName: CLICKSTREAM_ENGAGEMENT_PAGE_SCREEN_VIEW,
+    spName: CLICKSTREAM_ENGAGEMENT_PAGE_SCREEN_VIEW_SP,
     type: 'sp',
     scheduleRefresh: 'true',
   },
   {
-    viewName: CLICKSTREAM_RETENTION_DAU_WAU_SP,
+    viewName: CLICKSTREAM_RETENTION_DAU_WAU,
+    spName: CLICKSTREAM_RETENTION_DAU_WAU_SP,
     type: 'sp',
     scheduleRefresh: 'true',
   },
   {
-    viewName: CLICKSTREAM_RETENTION_EVENT_OVERTIME_SP,
+    viewName: CLICKSTREAM_RETENTION_EVENT_OVERTIME,
+    spName: CLICKSTREAM_RETENTION_EVENT_OVERTIME_SP,
     type: 'sp',
     scheduleRefresh: 'true',
   },
   {
-    viewName: CLICKSTREAM_RETENTION_USER_NEW_RETURN_SP,
+    viewName: CLICKSTREAM_RETENTION_USER_NEW_RETURN,
+    spName: CLICKSTREAM_RETENTION_USER_NEW_RETURN_SP,
     type: 'sp',
     scheduleRefresh: 'true',
   },
   {
-    viewName: CLICKSTREAM_DEVICE_CRASH_RATE_SP,
+    viewName: CLICKSTREAM_DEVICE_CRASH_RATE,
+    spName: CLICKSTREAM_DEVICE_CRASH_RATE_SP,
     type: 'sp',
     scheduleRefresh: 'true',
   },
