+++ src/analytics/private/sqls/redshift/dashboard/clickstream_acquisition_day_traffic_source_user_sp.sql
@@ -1,14 +1,14 @@
-CREATE OR REPLACE PROCEDURE {{database_name}}.{{schema}}.clickstream_acquisition_day_traffic_source_user_sp(day date) 
+CREATE OR REPLACE PROCEDURE {{database_name}}.{{schema}}.{{spName}}(day date) 
 LANGUAGE plpgsql
 AS $$ 
 DECLARE 
 
 BEGIN
 
-  DELETE FROM {{database_name}}.{{schema}}.clickstream_acquisition_day_traffic_source_user where event_date = day;
+  DELETE FROM {{database_name}}.{{schema}}.{{viewName}} where event_date = day;
 
   -- first_traffic_source
-  INSERT INTO {{database_name}}.{{schema}}.clickstream_acquisition_day_traffic_source_user (
+  INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
     event_date, 
     platform, 
     aggregation_type,
@@ -28,7 +28,7 @@ BEGIN
   ;
 
   -- first_traffic_source/medium
-  INSERT INTO {{database_name}}.{{schema}}.clickstream_acquisition_day_traffic_source_user (
+  INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
     event_date, 
     platform, 
     aggregation_type,
@@ -48,7 +48,7 @@ BEGIN
   ;
 
   -- first_traffic_medium
-  INSERT INTO {{database_name}}.{{schema}}.clickstream_acquisition_day_traffic_source_user (
+  INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
     event_date, 
     platform, 
     aggregation_type,
@@ -68,7 +68,7 @@ BEGIN
   ;
 
   -- first_traffic_campaign
-  INSERT INTO {{database_name}}.{{schema}}.clickstream_acquisition_day_traffic_source_user (
+  INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
     event_date, 
     platform, 
     aggregation_type,
@@ -88,7 +88,7 @@ BEGIN
   ;
 
   -- first_traffic_clid_platform
-  INSERT INTO {{database_name}}.{{schema}}.clickstream_acquisition_day_traffic_source_user (
+  INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
     event_date, 
     platform, 
     aggregation_type,
@@ -108,7 +108,7 @@ BEGIN
   ;
 
   -- first_traffic_channel_group,
-  INSERT INTO {{database_name}}.{{schema}}.clickstream_acquisition_day_traffic_source_user (
+  INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
     event_date, 
     platform, 
     aggregation_type,
@@ -128,7 +128,7 @@ BEGIN
   ;
 
   -- first_app_install_source
-  INSERT INTO {{database_name}}.{{schema}}.clickstream_acquisition_day_traffic_source_user (
+  INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
     event_date, 
     platform, 
     aggregation_type,
@@ -148,7 +148,7 @@ BEGIN
   ;
 
   -- session_source
-  INSERT INTO {{database_name}}.{{schema}}.clickstream_acquisition_day_traffic_source_user (
+  INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
     event_date, 
     platform, 
     aggregation_type,
@@ -167,7 +167,7 @@ BEGIN
   ;
 
   -- session_medium
-  INSERT INTO {{database_name}}.{{schema}}.clickstream_acquisition_day_traffic_source_user (
+  INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
     event_date, 
     platform, 
     aggregation_type,
@@ -186,7 +186,7 @@ BEGIN
   ;
 
   -- session_source/medium
-  INSERT INTO {{database_name}}.{{schema}}.clickstream_acquisition_day_traffic_source_user (
+  INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
     event_date, 
     platform, 
     aggregation_type,
@@ -205,7 +205,7 @@ BEGIN
   ;
 
   -- session_campaign
-  INSERT INTO {{database_name}}.{{schema}}.clickstream_acquisition_day_traffic_source_user (
+  INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
     event_date, 
     platform, 
     aggregation_type,
@@ -223,7 +223,7 @@ BEGIN
   group by 1,2,3,4,5
   ;
   -- session_clid_platform
-  INSERT INTO {{database_name}}.{{schema}}.clickstream_acquisition_day_traffic_source_user (
+  INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
     event_date, 
     platform, 
     aggregation_type,
@@ -241,7 +241,7 @@ BEGIN
   group by 1,2,3,4,5
   ;
   -- session_channel_group
-  INSERT INTO {{database_name}}.{{schema}}.clickstream_acquisition_day_traffic_source_user (
+  INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
     event_date, 
     platform, 
     aggregation_type,
@@ -260,7 +260,7 @@ BEGIN
   ;
 
 EXCEPTION WHEN OTHERS THEN
-    call {{database_name}}.{{schema}}.sp_clickstream_log_non_atomic('clickstream_acquisition_day_traffic_source_user', 'error', 'error message:' || SQLERRM);
+    call {{database_name}}.{{schema}}.sp_clickstream_log_non_atomic('{{viewName}}', 'error', 'error message:' || SQLERRM);
     RAISE INFO 'error message: %', SQLERRM;
 END;
 $$;
