+++ src/analytics/private/sqls/redshift/dashboard/clickstream_acquisition_day_user_acquisition.sql
@@ -1,4 +1,4 @@
-CREATE TABLE IF NOT EXISTS {{database_name}}.{{schema}}.clickstream_acquisition_day_user_acquisition (
+CREATE TABLE IF NOT EXISTS {{database_name}}.{{schema}}.{{viewName}} (
     event_date date,
     aggregation_type varchar(255),
     aggregation_dim varchar(65535),
@@ -9,7 +9,8 @@ CREATE TABLE IF NOT EXISTS {{database_name}}.{{schema}}.clickstream_acquisition_
     engagement_session_cnt bigint,
     engagement_rate decimal,
     avg_user_engagement_time_minutes double precision,
-    event_cnt bigint
+    event_cnt bigint,
+    update_timestamp timestamp DEFAULT getdate()
 )
 BACKUP YES
 SORTKEY(event_date)
