+++ src/analytics/private/sqls/redshift/dashboard/clickstream_device_crash_rate.sql
@@ -1,9 +1,10 @@
-CREATE TABLE IF NOT EXISTS {{database_name}}.{{schema}}.clickstream_device_crash_rate(
+CREATE TABLE IF NOT EXISTS {{database_name}}.{{schema}}.{{viewName}}(
     event_date date,
     platform varchar(255),
     app_version varchar(255),
     merged_user_id varchar(255),
-    crashed_user_id varchar(255)
+    crashed_user_id varchar(255),
+    update_timestamp timestamp DEFAULT getdate()
 )
 BACKUP YES
 SORTKEY(event_date)
