+++ src/analytics/private/sqls/redshift/dashboard/clickstream_device_crash_rate_sp.sql
@@ -1,13 +1,13 @@
-CREATE OR REPLACE PROCEDURE {{database_name}}.{{schema}}.clickstream_device_crash_rate_sp(day date) 
+CREATE OR REPLACE PROCEDURE {{database_name}}.{{schema}}.{{spName}}(day date) 
  LANGUAGE plpgsql
 AS $$ 
 DECLARE 
 
 BEGIN
 
-DELETE FROM {{database_name}}.{{schema}}.clickstream_device_crash_rate where event_date = day;
+DELETE FROM {{database_name}}.{{schema}}.{{viewName}} where event_date = day;
 
-INSERT INTO {{database_name}}.{{schema}}.clickstream_device_crash_rate (
+INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
   event_date, 
   platform,
   app_version,
@@ -26,7 +26,7 @@ group by 1, 2, 3, 4, 5
 ;
 
 EXCEPTION WHEN OTHERS THEN
-    call {{database_name}}.{{schema}}.sp_clickstream_log_non_atomic('clickstream_device_crash_rate', 'error', 'error message:' || SQLERRM);
+    call {{database_name}}.{{schema}}.sp_clickstream_log_non_atomic('{{viewName}}', 'error', 'error message:' || SQLERRM);
     RAISE INFO 'error message: %', SQLERRM;
 END;      
 $$
