+++ src/analytics/private/sqls/redshift/dashboard/clickstream_engagement_day_user_view.sql
@@ -1,8 +1,9 @@
-CREATE TABLE IF NOT EXISTS {{database_name}}.{{schema}}.clickstream_engagement_day_user_view (
+CREATE TABLE IF NOT EXISTS {{database_name}}.{{schema}}.{{viewName}} (
     event_date date,
     platform varchar(255),
     event_cnt bigint,
-    view_cnt bigint
+    view_cnt bigint,
+    update_timestamp timestamp DEFAULT getdate()
 )
 BACKUP YES
 SORTKEY(event_date)
