+++ src/analytics/private/sqls/redshift/dashboard/clickstream_engagement_day_user_view_sp.sql
@@ -1,13 +1,13 @@
-CREATE OR REPLACE PROCEDURE {{database_name}}.{{schema}}.clickstream_engagement_day_user_view_sp (day date) 
+CREATE OR REPLACE PROCEDURE {{database_name}}.{{schema}}.{{spName}} (day date) 
  LANGUAGE plpgsql
 AS $$ 
 DECLARE 
 
 BEGIN
 
-    DELETE FROM {{database_name}}.{{schema}}.clickstream_engagement_day_user_view where event_date = day;
+    DELETE FROM {{database_name}}.{{schema}}.{{viewName}} where event_date = day;
 
-    INSERT INTO {{database_name}}.{{schema}}.clickstream_engagement_day_user_view (
+    INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
         event_date,
         platform,
         event_cnt,
@@ -24,7 +24,7 @@ BEGIN
     ;
 
 EXCEPTION WHEN OTHERS THEN
-    call {{database_name}}.{{schema}}.sp_clickstream_log_non_atomic('clickstream_engagement_day_user_view', 'error', 'error message:' || SQLERRM);
+    call {{database_name}}.{{schema}}.sp_clickstream_log_non_atomic('{{viewName}}', 'error', 'error message:' || SQLERRM);
     RAISE INFO 'error message: %', SQLERRM;
 END;      
 $$
