+++ src/analytics/private/sqls/redshift/dashboard/clickstream_engagement_entrance.sql
@@ -1,11 +1,11 @@
-CREATE TABLE IF NOT EXISTS {{database_name}}.{{schema}}.clickstream_engagement_entrance (
+CREATE TABLE IF NOT EXISTS {{database_name}}.{{schema}}.{{viewName}} (
     event_date date,
     platform varchar(255),
     aggregation_type varchar(255),
     aggregation_dim varchar(65535),
-    entrance_cnt bigint
+    entrance_cnt bigint,
+    update_timestamp timestamp DEFAULT getdate()
 )
 BACKUP YES
-DISTSTYLE EVEN
 SORTKEY(event_date)
 ;
