+++ src/analytics/private/sqls/redshift/dashboard/clickstream_engagement_entrance_sp.sql
@@ -1,13 +1,13 @@
-CREATE OR REPLACE PROCEDURE {{database_name}}.{{schema}}.clickstream_engagement_entrance_sp (day date) 
+CREATE OR REPLACE PROCEDURE {{database_name}}.{{schema}}.{{spName}} (day date) 
  LANGUAGE plpgsql
 AS $$ 
 DECLARE 
 
 BEGIN
 
-    DELETE FROM {{database_name}}.{{schema}}.clickstream_engagement_entrance where event_date = day;
+    DELETE FROM {{database_name}}.{{schema}}.{{viewName}} where event_date = day;
 
-    INSERT INTO {{database_name}}.{{schema}}.clickstream_engagement_entrance (
+    INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
         event_date,
         platform,
         aggregation_type,
@@ -29,7 +29,7 @@ BEGIN
     group by 1, 2, 3, 4
     ;
 
-    INSERT INTO {{database_name}}.{{schema}}.clickstream_engagement_entrance (
+    INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
         event_date,
         platform,
         aggregation_type,
@@ -51,7 +51,7 @@ BEGIN
     group by 1, 2, 3, 4
     ;
 
-    INSERT INTO {{database_name}}.{{schema}}.clickstream_engagement_entrance (
+    INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
         event_date,
         platform,
         aggregation_type,
@@ -73,7 +73,7 @@ BEGIN
     group by 1, 2, 3, 4
     ;
 
-    INSERT INTO {{database_name}}.{{schema}}.clickstream_engagement_entrance (
+    INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
         event_date,
         platform,
         aggregation_type,
@@ -96,7 +96,7 @@ BEGIN
     ;
 
 EXCEPTION WHEN OTHERS THEN
-    call {{database_name}}.{{schema}}.sp_clickstream_log_non_atomic('clickstream_engagement_entrance', 'error', 'error message:' || SQLERRM);
+    call {{database_name}}.{{schema}}.sp_clickstream_log_non_atomic('{{viewName}}', 'error', 'error message:' || SQLERRM);
     RAISE INFO 'error message: %', SQLERRM;
 END;      
 $$
