+++ src/analytics/private/sqls/redshift/dashboard/clickstream_engagement_exit_sp.sql
@@ -1,13 +1,13 @@
-CREATE OR REPLACE PROCEDURE {{database_name}}.{{schema}}.clickstream_engagement_exit_sp (day date) 
+CREATE OR REPLACE PROCEDURE {{database_name}}.{{schema}}.{{spName}} (day date) 
  LANGUAGE plpgsql
 AS $$ 
 DECLARE 
 
 BEGIN
 
-    DELETE FROM {{database_name}}.{{schema}}.clickstream_engagement_exit where event_date = day;
+    DELETE FROM {{database_name}}.{{schema}}.{{viewName}} where event_date = day;
 
-    INSERT INTO {{database_name}}.{{schema}}.clickstream_engagement_exit (
+    INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
         event_date,
         platform,
         aggregation_type,
@@ -46,7 +46,7 @@ BEGIN
       1,2,3,4;
 
 
-    INSERT INTO {{database_name}}.{{schema}}.clickstream_engagement_exit (
+    INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
         event_date,
         platform,
         aggregation_type,
@@ -85,7 +85,7 @@ BEGIN
       1,2,3,4;
 
 
-    INSERT INTO {{database_name}}.{{schema}}.clickstream_engagement_exit (
+    INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
         event_date,
         platform,
         aggregation_type,
@@ -124,7 +124,7 @@ BEGIN
       1,2,3,4;
 
 
-    INSERT INTO {{database_name}}.{{schema}}.clickstream_engagement_exit (
+    INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
         event_date,
         platform,
         aggregation_type,
@@ -163,7 +163,7 @@ BEGIN
       1,2,3,4;
 
 EXCEPTION WHEN OTHERS THEN
-    call {{database_name}}.{{schema}}.sp_clickstream_log_non_atomic('clickstream_engagement_exit', 'error', 'error message:' || SQLERRM);
+    call {{database_name}}.{{schema}}.sp_clickstream_log_non_atomic('{{viewName}}', 'error', 'error message:' || SQLERRM);
     RAISE INFO 'error message: %', SQLERRM;
 END;      
 $$
