+++ src/analytics/private/sqls/redshift/dashboard/clickstream_engagement_page_screen_view_detail.sql
@@ -1,13 +1,13 @@
-CREATE TABLE IF NOT EXISTS {{database_name}}.{{schema}}.clickstream_engagement_page_screen_view_detail (
+CREATE TABLE IF NOT EXISTS {{database_name}}.{{schema}}.{{viewName}} (
     event_date date,
     platform varchar(255),
     aggregation_type varchar(255),
     aggregation_dim varchar(65535),
     user_id varchar(255),
     user_engagement_time_minutes double precision,
-    event_id varchar(255)
+    event_id varchar(255),
+    update_timestamp timestamp DEFAULT getdate()
 )
 BACKUP YES
-DISTSTYLE EVEN
 SORTKEY(event_date)
 ;
