+++ src/analytics/private/sqls/redshift/dashboard/clickstream_engagement_page_screen_view_detail_sp.sql
@@ -1,13 +1,13 @@
-CREATE OR REPLACE PROCEDURE {{database_name}}.{{schema}}.clickstream_engagement_page_screen_view_detail_sp (day date) 
+CREATE OR REPLACE PROCEDURE {{database_name}}.{{schema}}.{{spName}} (day date) 
  LANGUAGE plpgsql
 AS $$ 
 DECLARE 
 
 BEGIN
 
-    DELETE FROM {{database_name}}.{{schema}}.clickstream_engagement_page_screen_view_detail where event_date = day;
+    DELETE FROM {{database_name}}.{{schema}}.{{viewName}} where event_date = day;
 
-    INSERT INTO {{database_name}}.{{schema}}.clickstream_engagement_page_screen_view_detail (
+    INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
         event_date,
         platform,
         aggregation_type,
@@ -30,7 +30,7 @@ BEGIN
     and event_name = '_page_view'
     ;
 
-    INSERT INTO {{database_name}}.{{schema}}.clickstream_engagement_page_screen_view_detail (
+    INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
         event_date,
         platform,
         aggregation_type,
@@ -53,7 +53,7 @@ BEGIN
     and event_name = '_page_view'
     ;
 
-    INSERT INTO {{database_name}}.{{schema}}.clickstream_engagement_page_screen_view_detail (
+    INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
         event_date,
         platform,
         aggregation_type,
@@ -76,7 +76,7 @@ BEGIN
     and event_name = '_screen_view'
     ;
 
-    INSERT INTO {{database_name}}.{{schema}}.clickstream_engagement_page_screen_view_detail (
+    INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (
         event_date,
         platform,
         aggregation_type,
@@ -100,7 +100,7 @@ BEGIN
     ;
 
 EXCEPTION WHEN OTHERS THEN
-    call {{database_name}}.{{schema}}.sp_clickstream_log_non_atomic('clickstream_engagement_page_screen_view_detail', 'error', 'error message:' || SQLERRM);
+    call {{database_name}}.{{schema}}.sp_clickstream_log_non_atomic('{{viewName}}', 'error', 'error message:' || SQLERRM);
     RAISE INFO 'error message: %', SQLERRM;
 END;      
 $$
