+++ src/analytics/private/sqls/redshift/dashboard/clickstream_lifecycle_view_v2.sql
@@ -4,9 +4,8 @@ AUTO REFRESH NO
 AS
 select 
   user_pseudo_id, 
-  DATE_TRUNC('day', CONVERT_TIMEZONE('{{timezone}}', event_timestamp)) as time_period,
   DATE_TRUNC('week', CONVERT_TIMEZONE('{{timezone}}', event_timestamp)) as time_period_week
 from {{database_name}}.{{schema}}.event_v2
 where event_name = '_session_start' 
-group by 1,2,3
+group by 1,2
 ;
