+++ src/analytics/private/sqls/redshift/dashboard/clickstream_retention_event_overtime.sql
@@ -1,7 +1,8 @@
-CREATE TABLE IF NOT EXISTS {{database_name}}.{{schema}}.clickstream_retention_event_overtime(
+CREATE TABLE IF NOT EXISTS {{database_name}}.{{schema}}.{{viewName}}(
     event_date date,
     platform varchar(255),
-    event_cnt bigint
+    event_cnt bigint,
+    update_timestamp timestamp DEFAULT getdate()
 )
 BACKUP YES
 SORTKEY(event_date)
