+++ src/analytics/private/sqls/redshift/dashboard/clickstream_retention_user_new_return_sp.sql
@@ -1,17 +1,17 @@
-CREATE OR REPLACE PROCEDURE {{database_name}}.{{schema}}.clickstream_retention_user_new_return_sp(day date) 
+CREATE OR REPLACE PROCEDURE {{database_name}}.{{schema}}.{{spName}}(day date) 
  LANGUAGE plpgsql
 AS $$ 
 DECLARE 
 
 BEGIN
 
-DELETE FROM {{database_name}}.{{schema}}.clickstream_retention_user_new_return where event_date = day;
+DELETE FROM {{database_name}}.{{schema}}.{{viewName}} where event_date = day;
 
-INSERT INTO {{database_name}}.{{schema}}.clickstream_retention_user_new_return (event_date, platform, user_type, user_cnt)
+INSERT INTO {{database_name}}.{{schema}}.{{viewName}} (event_date, platform, user_type, user_cnt)
 select 
   event_date,
   platform,
-  case when day = first_visit_date then 'NEW' else 'RETURN' end as user_type,
+  case when day = first_visit_date then 'NEW' else 'RETURNING' end as user_type,
   count(distinct merged_user_id) as user_cnt
 from {{database_name}}.{{schema}}.{{baseView}}
 where event_date = day
@@ -19,7 +19,7 @@ group by 1,2,3
 ;
 
 EXCEPTION WHEN OTHERS THEN
-    call {{database_name}}.{{schema}}.sp_clickstream_log_non_atomic('clickstream_retention_user_new_return', 'error', 'error message:' || SQLERRM);
+    call {{database_name}}.{{schema}}.sp_clickstream_log_non_atomic('{{viewName}}', 'error', 'error message:' || SQLERRM);
     RAISE INFO 'error message: %', SQLERRM;
 END;      
 $$
